Set Tile Size to 1x1 and made a map for it
Tile size has been set to 1x1. To account for this, I made (and somehow got working) a new map with a 1x1 tile size to reflect the change.
Although, there seems to be a good bit of lag on the 1x1 map. I may have to edit this later.